fix(step3): random 함수 테스트 코드 분리
